UBUNTU: [Config] Compile the rtc-pl031 driver builtin on the highbank kernel flavour
The rtc-pl031 driver is used for the RTC on the highbank
platform. Presently it is built as a module in Precise and
Quantal, making the driver unavailable in early boot, yielding
the following error on both releases:

./drivers/rtc/hctosys.c: unable to open rtc device (rtc0)

This behaviour prevents the ability to accurately check the
last-mounted time on the root filesystem.

BugLink: https://bugs.launchpad.net/bugs/1035110
Signed-off-by: Dann Frazier <dann.frazier@canonical.com>
Signed-off-by: Chris Van Hoof <vanhoof@canonical.com>
Cc: Ike Pan <ike.pan@canonical.com>
Acked-by: Brad Figg <brad.figg@canonical.com>
Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>