Revert "UBUNTU: SAUCE: Force xsave off on older Xen hypervisors"
This reverts commit 85c46aa8b6c17d1a9360559d22f4683b1a38e33e.

It turned out that since 2.6.39 the Xen code in the kernel would
not try to set OSXSAVE in CR4 to find out whether the feature is
supported. Instead it relies on OSXSAVE being set by the HV in
the cpuid bits (which will only happen if supported).

BugLink: http://bugs.launchpad.net/bugs/1044550

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Colin King <colin.king@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>